pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["Joel Dumoulin <joel.dumoulin@idiap.ch>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
click = "^8.0.3"
advent-of-code-data = "^1.1.0"
attrs = "^21.2.0"
numpy = "^1.21.4"
networkx = "^2.6.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
